# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Oil \u0026 Gas Service Booking API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8082",
    "basePath": "/",
    "paths": {
...
        "/companies/{company_id}/services": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "company-services"
                ],
                "summary": "Получить услуги компании",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Company ID",
                        "name": "company_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/oil-gas-service-booking_internal_models.CompanyService"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/company-services": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "company-services"
                ],
                "summary": "Получить все связи компаний и услуг",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/oil-gas-service-booking_internal_models.CompanyService"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "company-services"
                ],
                "summary": "Создать связь компании и услуги",
                "parameters": [
                    {
                        "description": "Данные связи компании и услуги",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_http-server_handlers.CompanyServiceCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/oil-gas-service-booking_internal_models.CompanyService"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/company-services/my": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "company-services"
                ],
                "summary": "Получить связи услуг только моих компаний",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/oil-gas-service-booking_internal_models.CompanyService"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/company-services/{id}": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "company-services"
                ],
                "summary": "Получить связь компании и услуги по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "CompanyService ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/oil-gas-service-booking_internal_models.CompanyService"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "company-services"
                ],
                "summary": "Обновить связь компании и услуги",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "CompanyService ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные для обновления",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/oil-gas-service-booking_internal_models.CompanyService"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/oil-gas-service-booking_internal_models.CompanyService"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "tags": [
                    "company-services"
                ],
                "summary": "Удалить связь компании и услуги",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "CompanyService ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/services/{service_id}/companies": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "company-services"
                ],
                "summary": "Получить компании, предоставляющие услугу",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Service ID",
                        "name": "service_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/oil-gas-service-booking_internal_models.CompanyService"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_http-server_handlers.CompanyServiceCreateRequest": {
            "type": "object",
            "properties": {
                "company_id": {
                    "type": "integer"
                },
                "service_id": {
                    "type": "integer"
                }
            }
        },
...
        "oil-gas-service-booking_internal_models.Booking": {
            "type": "object",
            "properties": {
                "bookingID": {
                    "type": "integer"
                },
                "bookingServices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/oil-gas-service-booking_internal_models.BookingService"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/oil-gas-service-booking_internal_models.User"
                },
                "userID": {
                    "type": "integer"
                }
            }
        },
        "oil-gas-service-booking_internal_models.BookingService": {
            "type": "object",
            "properties": {
                "booking": {
                    "$ref": "#/definitions/oil-gas-service-booking_internal_models.Booking"
                },
                "bookingID": {
                    "type": "integer"
                },
                "bookingServiceID": {
                    "type": "integer"
                },
                "companyService": {
                    "$ref": "#/definitions/oil-gas-service-booking_internal_models.CompanyService"
                },
                "companyServiceID": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "oil-gas-service-booking_internal_models.Company": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "companyID": {
                    "type": "integer"
                },
                "companyServices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/oil-gas-service-booking_internal_models.CompanyService"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/oil-gas-service-booking_internal_models.User"
                },
                "userID": {
                    "description": "владелец",
                    "type": "integer"
                }
            }
        },
        "oil-gas-service-booking_internal_models.CompanyService": {
            "type": "object",
            "properties": {
                "bookingServices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/oil-gas-service-booking_internal_models.BookingService"
                    }
                },
                "company": {
                    "$ref": "#/definitions/oil-gas-service-booking_internal_models.Company"
                },
                "companyID": {
                    "type": "integer"
                },
                "companyServiceID": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "price": {
                    "description": "цена конкретной услуги у компании",
                    "type": "number"
                },
                "service": {
                    "$ref": "#/definitions/oil-gas-service-booking_internal_models.Service"
                },
                "serviceID": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "oil-gas-service-booking_internal_models.Service": {
            "type": "object",
            "properties": {
                "companyServices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/oil-gas-service-booking_internal_models.CompanyService"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "serviceID": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "oil-gas-service-booking_internal_models.User": {
            "type": "object",
            "properties": {
                "bookings": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/oil-gas-service-booking_internal_models.Booking"
                    }
                },
                "companies": {
                    "description": "Добавляем обратную связь",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/oil-gas-service-booking_internal_models.Company"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userID": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}